#10900 This commit includes: - Better exception handling when an assertion is null and/or attributes can not be got from the assertion - Docs updates - DotSamlException and AttributesNotFoundException were moved to com.dotcms.plugin.saml.v3.exception
