+++ .gitignore
@@ -1,0 +1,5 @@
+build/
+out/
+*.iml
+.idea/
+.DS_Store*
