+++ src/com/dotcms/plugin/saml/v3/handler/SOAPArtifactAssertionResolverHandlerImpl.java
@@ -1,7 +1,7 @@
 package com.dotcms.plugin.saml.v3.handler;
 
 import com.dotcms.plugin.saml.v3.DotSamlConstants;
-import com.dotcms.plugin.saml.v3.DotSamlException;
+import com.dotcms.plugin.saml.v3.exception.DotSamlException;
 import com.dotcms.plugin.saml.v3.HttpClient;
 import com.dotcms.plugin.saml.v3.SiteConfigurationResolver;
 import com.dotcms.plugin.saml.v3.config.Configuration;
